+++ packages/schema/types/user.ts
@@ -1,4 +1,5 @@
 import { extendType, objectType } from "@nexus/schema";
+import { isAuthenticated } from "../rules";
 
 export const User = objectType({
   name: "User",
@@ -11,14 +12,14 @@ export const User = objectType({
 export const UserQuery = extendType({
   type: "Query",
   definition(t) {
-    t.crud.user();
-    t.crud.users();
+    t.crud.user({ shield: isAuthenticated() });
+    t.crud.users({ shield: isAuthenticated() });
   },
 });
 
 export const UserMutation = extendType({
   type: "Mutation",
   definition(t) {
-    t.crud.deleteOneUser();
+    t.crud.deleteOneUser({ shield: isAuthenticated() });
   },
 });